requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
apptools==5.3.0
asgiref==3.8.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
av==12.3.0
backoff==2.2.1
backports.strenum==1.3.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bokeh==3.4.1
build==1.2.1
cachetools==5.4.0
Cartopy==0.23.0
certifi==2024.7.4
chainlit==1.1.401
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
databricks-sdk==0.29.0
dataclasses-json==0.5.14
datasets==2.19.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docopt==0.6.2
einops==0.8.0
entrypoints==0.4
envisage==7.0.3
eva-decord==0.6.1
exceptiongroup==1.2.2
fastapi==0.110.3
fastapi-socketio==0.0.10
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
giskard==2.11.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
griffe==0.48.0
grpcio==1.65.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
inquirerpy==0.3.4
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.96
Lazify==0.4.0
literalai==0.0.607
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
mayavi==4.8.2
mdurl==0.1.2
mixpanel==4.10.1
mlflow-skinny==2.15.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.1.9
onnxruntime==1.18.1
openai==1.27.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pfzy==0.3.4
pillow==10.4.0
portalocker==2.10.1
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.4
pulsar-client==3.5.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydicom==2.4.4
pyface==8.0.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.1.0
pysbd==0.3.4
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.3
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.7
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.11.4
sentry-sdk==2.12.0
shapely==2.0.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.37.2
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.6.7
tomli==2.0.1
tqdm==4.66.4
traitsui==8.0.0
transformers==4.43.3
triton==3.0.0
typer==0.12.3
types-regex==2024.7.24.20240726
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.25.0
uvloop==0.19.0
vtk==9.3.1
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.27.post2
xxhash==3.4.1
xyzservices==2024.6.0
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
